requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
Django==5.0.3
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-htmx==1.18.0
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django_csp==3.8
filelock==3.16.1
identify==2.6.1
isort==5.13.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.40
pillow==10.3.0
platformdirs==4.2.2
pre_commit==4.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.4.4
tomli==2.0.2
typing_extensions==4.12.2
virtualenv==20.26.6
